+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - New workflow for app updates, including in app changelog for the new release and an automated update installer which deletes the older portable exe and replaces it with the new one
 - Fixed an issue with re-authentication timeout config not being saved to the app settings file.
 - Fixed an issue with animations smoothness when searching tokens
+- Added a clear search button in the searchbar
 
 # v0.5.2
 - Fixed an issue blocking icons from loading when app protection is enabled